overhead-people-detection/requirements.txt
altair==5.0.1
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
cryptography @ file:///D:/bld/cryptography_1666563584497/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython @ file:///D:/bld/cython_1659102065258/work
decorator==5.1.1
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools @ file:///D:/bld/fonttools_1666390069478/work
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.3
jsonschema==4.17.3
kiwisolver @ file:///D:/bld/kiwisolver_1657953189205/work
markdown-it-py==2.2.0
MarkupSafe==2.1.4
matplotlib @ file:///D:/bld/matplotlib-suite_1661439973878/work
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
munkres==1.1.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
opencv-python==4.9.0.80
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==1.3.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.4
pyarrow==12.0.1
pycocotools @ file:///D:/bld/pycocotools_1661485949547/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
pyrsistent==0.19.3
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
requests @ file:///C:/ci/requests_1657735288441/work
rich==13.7.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm @ file:///C:/b/abs_0axbz66qik/croots/recipe/tqdm_1664392691071/work
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
tzdata==2023.4
tzlocal==4.3.1
unicodedata2 @ file:///D:/bld/unicodedata2_1649112131705/work
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
validators==0.20.0
watchdog==3.0.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
zipp==3.15.0
